🔧 Split components
